+++ lib/aead/AbstractBufferAeadCcm.ts
@@ -1,18 +1,16 @@
-import AbstractBufferAead from '../../aead/AbstractBufferAead';
-import BufferAeadType from '../../types/BufferAeadType';
+import AbstractBufferAead from './AbstractBufferAead';
+import BufferAeadType from '../types/BufferAeadType';
 import crypto from 'crypto';
-import DecryptionInput from '../../types/DecryptionInput';
-import EncryptionInput from '../../types/EncryptionInput';
-import EncryptionOutput from '../../types/EncryptionOutput';
+import DecryptionInput from '../types/DecryptionInput';
+import EncryptionInput from '../types/EncryptionInput';
+import EncryptionOutput from '../types/EncryptionOutput';
 
-abstract class AbstractBufferAeadAesGcm extends AbstractBufferAead {
+const authTagLength = 16;
+
+abstract class AbstractBufferAeadCcm extends AbstractBufferAead {
   private readonly algorithm: BufferAeadType;
 
-  public constructor (algorithm: BufferAeadType) {
-    // eslint-disable-next-line prefer-named-capture-group
-    const keyLength = Number.parseInt(algorithm.replace(/^aes-(128|192|256)-gcm$/u, '$1'), 10) / 8;
-    const nonceLength = 12;
-
+  public constructor ({ algorithm, keyLength, nonceLength }: { algorithm: BufferAeadType; keyLength: number; nonceLength: number }) {
     super({ keyLength, nonceLength });
     this.algorithm = algorithm;
   }
@@ -22,8 +20,8 @@ abstract class AbstractBufferAeadAesGcm extends AbstractBufferAead {
     const nonce = input.nonce ?? this.nonceGen();
     const additionalData = input.additionalData ?? Buffer.alloc(0);
 
-    const encrypter = crypto.createCipheriv(this.algorithm, key, nonce) as crypto.CipherGCM;
-    encrypter.setAAD(additionalData);
+    const encrypter = crypto.createCipheriv(this.algorithm, key, nonce, { authTagLength } as any) as crypto.CipherCCM;
+    encrypter.setAAD(additionalData, { plaintextLength: input.data.length });
 
     const ciphertext = Buffer.concat([ encrypter.update(input.data), encrypter.final() ]);
     const authTag = encrypter.getAuthTag();
@@ -35,8 +33,8 @@ abstract class AbstractBufferAeadAesGcm extends AbstractBufferAead {
     const { ciphertext, authTag, key, nonce } = input;
     const additionalData = input.additionalData ?? Buffer.alloc(0);
 
-    const decrypter = crypto.createDecipheriv(this.algorithm, key, nonce) as crypto.DecipherGCM;
-    decrypter.setAAD(additionalData);
+    const decrypter = crypto.createDecipheriv(this.algorithm, key, nonce, { authTagLength } as any) as crypto.DecipherCCM;
+    decrypter.setAAD(additionalData, { plaintextLength: ciphertext.length });
     decrypter.setAuthTag(authTag);
 
     try {
@@ -52,4 +50,4 @@ abstract class AbstractBufferAeadAesGcm extends AbstractBufferAead {
   }
 }
 
-export default AbstractBufferAeadAesGcm;
+export default AbstractBufferAeadCcm;
