+++ lib/aead/BufferAeadAes128Gcm.ts
@@ -1,4 +1,4 @@
-import AbstractBufferAeadAesGcm from './AbstractBufferAeadAesGcm';
+import AbstractBufferAeadAesGcm from './AbstractBufferAeadGcm';
 
 class BufferAeadAes128Gcm extends AbstractBufferAeadAesGcm {
   public constructor () {
