--- lib/aead/aesccm/BufferAeadAes128Ccm.ts
@@ -1,9 +1,0 @@
-import AbstractBufferAeadAesCcm from './AbstractBufferAeadAesCcm';
-
-class BufferAeadAes128Ccm extends AbstractBufferAeadAesCcm {
-  public constructor () {
-    super('aes-128-ccm');
-  }
-}
-
-export default BufferAeadAes128Ccm;
