--- lib/aead/aesccm/BufferAeadAes192Ccm.ts
@@ -1,9 +1,0 @@
-import AbstractBufferAeadAesCcm from './AbstractBufferAeadAesCcm';
-
-class BufferAeadAes192Ccm extends AbstractBufferAeadAesCcm {
-  public constructor () {
-    super('aes-192-ccm');
-  }
-}
-
-export default BufferAeadAes192Ccm;
