+++ lib/types/BufferAeadType.ts
@@ -1,5 +1,5 @@
 import crypto from 'crypto';
 
-type BufferAeadType = crypto.CipherGCMTypes | Exclude<crypto.CipherCCMTypes, 'chacha20-poly1305'> | 'aes-ctrhmac' | 'xchacha20-poly1305';
+type BufferAeadType = crypto.CipherGCMTypes | crypto.CipherCCMTypes | 'aes-ctrhmac' | 'xchacha20-poly1305';
 
 export default BufferAeadType;
